Nuevo Handler para las credenciales
