+++ Prueba.Tecnica.Web.API/Controllers/AuthController.cs
@@ -1,8 +1,10 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using MediatR;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity.Data;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
 using Prueba.Tecnica.Web.Application.Feature.Auth.Login.Command;
+using Prueba.Tecnica.Web.Application.Feature.Files.Queries;
 using Prueba.Tecnica.Web.Application.ResponseModels;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
@@ -13,18 +15,21 @@ namespace Prueba.Tecnica.Web.API.Controllers
     /// <summary>
     /// Controlador para la autenticacion de usuarios
     /// </summary>
-    [Route("api/[controller]")]
+    [Route("api/auth")]
     [ApiController]
     public class AuthController : ControllerBase
     {
-        private readonly IConfiguration _configuration;
+        private readonly IMediator _mediator;
+        private readonly ILogger<AuthController> _logger;
         /// <summary>
         /// 
         /// </summary>
-        /// <param name="configuration"></param>
-        public AuthController(IConfiguration configuration)
+        /// <param name="logger"></param>
+        /// <param name="mediator"></param>
+        public AuthController( ILogger<AuthController> logger, IMediator mediator)
         {
-            _configuration = configuration;
+            _logger = logger;
+            _mediator = mediator;
         }
 
 
@@ -38,8 +43,8 @@ namespace Prueba.Tecnica.Web.API.Controllers
         ///
         ///     POST 
         ///     {
-        ///        "username": "usuario",
-        ///        "password": "password"
+        ///        "username": "prueba",
+        ///        "password": "tecnica"
         ///     }
         ///
         /// </remarks>
@@ -48,37 +53,20 @@ namespace Prueba.Tecnica.Web.API.Controllers
         [HttpPost("login")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        public IActionResult Login([FromBody] LoginCommand request)
+        public async Task<IActionResult> Login([FromBody] LoginCommand request)
         {
-        
-            if (request.Username == "usuario" && request.Password == "password")
+            _logger.LogInformation($"Intentando iniciar sesion con el usuario: {request.Username}");
+            if (request.Username == "prueba" && request.Password == "tecnica")
             {
-                var expires = DateTime.Now.AddHours(1);
-                var token = GenerateJwtToken(request.Username, expires);
+                var response = await _mediator.Send(request);
 
-                var response = new LoginResponse
-                {
-                    Token = token,
-                    Expiration = expires
-                };
+                _logger.LogInformation($"Inicio de sesion exitoso para el usuario: {request.Username}");
                 return Ok(ApiResponse<LoginResponse>.CreateSuccessResponse(response));
             }
-
+            
             return Unauthorized();
         }
 
-        private string GenerateJwtToken(string username,DateTime expires)
-        {
-            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["Jwt:Key"]));
-            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-
-            var token = new JwtSecurityToken(
-                issuer: _configuration["Jwt:Issuer"],
-                claims: new[] { new Claim(ClaimTypes.Name, username) },
-                expires: expires,
-                signingCredentials: creds);
-
-            return new JwtSecurityTokenHandler().WriteToken(token);
-        }
+     
     }
 }
